// viasoft.licensing.customerlicensing-frontend/src/client/customer-licensing/model/licenseUsageInRealTimeTreeTableData.ts
/**
 * Viasoft.Licensing.CustomerLicensing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LicenseUsageInRealTimeTreeTableTypes } from './licenseUsageInRealTimeTreeTableTypes';


export interface LicenseUsageInRealTimeTreeTableData { 
    name?: string | null;
    description?: string | null;
    startTime?: Date;
    type?: LicenseUsageInRealTimeTreeTableTypes;
    readonly typeDescription?: string | null;
    total?: number;
    additional?: number;
    parent?: string | null;
    consumed?: number;
}