// viasoft.licensing.customerlicensing-frontend/src/client/customer-licensing/model/licenseUsageInRealTimeTreeTableTypes.ts
/**
 * Viasoft.Licensing.CustomerLicensing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type LicenseUsageInRealTimeTreeTableTypes = 'User' | 'BundledApp' | 'LooseApp' | 'Bundle';

export const LicenseUsageInRealTimeTreeTableTypes = {
    User: 'User' as LicenseUsageInRealTimeTreeTableTypes,
    BundledApp: 'BundledApp' as LicenseUsageInRealTimeTreeTableTypes,
    LooseApp: 'LooseApp' as LicenseUsageInRealTimeTreeTableTypes,
    Bundle: 'Bundle' as LicenseUsageInRealTimeTreeTableTypes
};